add: logout confirmation

+++ src/app/features/layout/layout.component.ts
@@ -1,5 +1,6 @@
 import { Component } from '@angular/core';
 import { Router } from '@angular/router';
+import { NzModalService } from 'ng-zorro-antd/modal';
 
 @Component({
   selector: 'app-layout',
@@ -8,11 +9,26 @@ import { Router } from '@angular/router';
 })
 export class LayoutComponent {
 
-  constructor (private router: Router) { }
+  constructor (
+    private router: Router,
+    private modal: NzModalService
+  ) { }
 
   logout() {
-    sessionStorage.clear();
-    this.router.navigate(['/login']);
+
+    this.modal.confirm({
+      nzTitle: 'ยืนยัน?',
+      nzContent: `ต้องการออกจากระบบ ใช่หรือไม่?`,
+      nzOkText: 'ใช่',
+      nzOkType: 'primary',
+      nzOkDanger: false,
+      nzOnOk: () => {
+        sessionStorage.clear();
+        this.router.navigate(['/login']);
+      },
+      nzCancelText: 'ไม่ใช่',
+      nzOnCancel: () => { }
+    });
   }
 
 }
